intermediate refactoring commit 2
